srcs/Managers_view_front/hr_mvp/requirements.txt
Django>=5,<6
# Optional extras for later:
# markdown
# openai>=1.0.0
requests>=2.0.0
python-dotenv>=1.0.0
